requirements.txt
aiogram==3.13.1
jinja2==3.1.4
requests==2.32.3
PyYAML==6.0.2
pandas==2.2.3
openpyxl==3.1.5

# Optional extractors (install later, after Python 3.12 is set and wheels available)
# trafilatura==1.12.2
# readability-lxml==0.8.1
# pdfminer.six==20231228
# pillow==10.4.0
# pymupdf==1.24.10
# paddleocr==2.9.1
# pytesseract==0.3.10


